First steps on TCP/UDP configuration rework; Nodes still need a lot of work
