+++ src/host/mod.rs
@@ -1,12 +1,10 @@
 mod config;
 #[allow(clippy::module_inception)]
 mod host;
+mod network_config;
 mod tcp;
-mod tcp_config;
 mod udp;
-mod udp_config;
 
 pub use crate::host::config::*;
 pub use crate::host::host::*;
-pub use crate::host::tcp_config::*;
-pub use crate::host::udp_config::*;
+pub use crate::host::network_config::*;
